requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
psutil>=5.9.0
GPUtil>=1.4.0
pynvml>=11.5.0
pystray>=0.19.0
Pillow>=10.0.0
WMI>=1.5.1
pywin32>=306
